Refactor a bunch of data manipulation methods to data.py
